+++ soundSource.h
@@ -9,7 +9,7 @@
 class soundSource{
 	
 	pa_simple *pa_connection;
-	uint8_t *pcm_buffer;
+	uint16_t *pcm_buffer;
 
 	void pa_server_info_callback();
 	void pa_context_state_callback();
@@ -17,7 +17,7 @@ class soundSource{
 	public:
 	soundSource();
 	void read_stream();
-	uint8_t* get_pcm_buffer();
+	uint16_t* get_pcm_buffer();
 	~soundSource();
 
 };
